requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
psycopg2-binary
pyjwt[crypto]
httpx
python-dotenv